Posters 2022 V1
This version has a single sigma for the judges a t distribution for their means. It seems reasonable. However, the width of the t distribution has a big effects and I think it makes sense to make it hierarchical so that the width is determined from the data. Maybe the thing that shouldn't be hierarchical is actually the df of the t distribution.